pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ecg-digitization"
version = "0.1.0"
description = "ECG Image Digitization - Extract time-series data from ECG paper printouts"
readme = "README.md"
license = {text = "BSD-3-Clause"}
requires-python = ">=3.10"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = [
    "ecg",
    "digitization",
    "medical-imaging",
    "deep-learning",
    "computer-vision",
    "signal-processing",
    "kaggle",
    "physionet"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
]

dependencies = [
    # Core ML/DL - these are in base image but listed for reference
    # "torch>=2.0.0",
    # "torchvision>=0.15.0",
    
    # Image Processing
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "albumentations>=1.3.0",
    
    # Signal Processing
    "wfdb>=4.1.0",  # PhysioNet WFDB library for ECG data
    "neurokit2>=0.2.0",  # ECG signal processing
    "biosppy>=1.0.0",  # Biosignal processing
    
    # Data handling
    # "pandas>=2.0.0",  # in base image
    # "numpy>=1.24.0",  # in base image
    "pyarrow>=14.0.0",
    
    # Model frameworks
    "transformers>=4.35.0",  # HuggingFace models
    "timm>=0.9.0",  # PyTorch Image Models
    "segmentation-models-pytorch>=0.3.0",
    
    # Experiment tracking
    "wandb>=0.16.0",
    "mlflow>=2.9.0",
    
    # Utilities
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "tqdm>=4.66.0",
    "loguru>=0.7.0",
    "rich>=13.0.0",
    
    # Visualization
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]
jupyter = [
    "jupyter>=1.0.0",
    "jupyterlab>=4.0.0",
    "ipywidgets>=8.0.0",
]

[project.urls]
"Homepage" = "https://www.kaggle.com/competitions/physionet-ecg-image-digitization"
"Repository" = "https://github.com/yourusername/ecg-digitization"

[project.scripts]
ecg-train = "ecg_digitization.train:main"
ecg-predict = "ecg_digitization.predict:main"
ecg-submit = "ecg_digitization.submit:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 100
target-version = ['py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | data
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["ecg_digitization"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
